python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# ML / CatVTON
torch>=2.2.0,<2.3.0
diffusers>=0.28.0,<0.33.0
transformers>=4.40.0
accelerate>=0.28.0
huggingface_hub>=0.22.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0
requests>=2.31.0
tqdm>=4.66.0
